feat(Tile/Dimmer): restart cycle if torch was disabled (#78)
* feat(tiles/dimmer): restart cycle when torch is disabled externally

Changes the DimmerSettingsTile to always start from the dimmest mode
if the toch is disabled. Previously it would continue from the last
mode, which was confusing to the user, since suddenly the tile acted
different when tapped.

* fix(Tiles/Dimmer): always display off state if torch is disabled

Fixes an issue, where if the torch was externally disabled, the mode
should continue to show the (now outdated) mode instead of off.

* fix: restructure settings tile logic

---------

Co-authored-by: Niko Diamadis <cyb3rko@pm.me>